+++ zeapp/app/src/main/java/de/berlindroid/zeapp/Config.kt
@@ -1,0 +1,4 @@
+package de.berlindroid.zeapp
+
+const val PAGE_WIDTH = 296
+const val PAGE_HEIGHT = 128
